feat(game-data)!: one record per entity — remove the gameContent namespace
`HEROES` entries now carry the gameplay sheet (bannerAction, defaultVirtues,
unlockableVirtues) alongside board identity, and the foe/adversary/companion
copies are gone in favour of the rosters that already covered them. The
`gameContent` namespace no longer exists.

Through v2 the same 14 heroes existed twice: identity in heroes.ts keyed by a
stable `id`, and the gameplay sheet in gameContent.HEROES keyed by display name
with no `id` at all. Nothing could join the two halves except by matching a
display string, and they could drift apart silently. The same held for foes,
adversaries and companions — each a thinner copy of a flat roster.

That was never a domain boundary. Those duplicates were the only things
colliding (Hero, HEROES, Foe, FOES, Adversary), and the namespace was the
workaround for the collision, recorded as such in src/index.ts. Removing the
duplicates removed the collisions, leaving the namespace nothing to protect —
so its last member, the kingdom virtues, is now flat in the new virtues.ts.

None of the removed exports had a consumer, here or in any app. The identity
half has many: apps/digital (hero pickers, player boards) and packages/board
(default roster, and portrait asset paths keyed by `id`) — so the merge keeps
the id-keyed roster and folds the sheet into it. Those consumers compile
unchanged; the new sheet fields are additive and optional.

Nothing was lost, and each claim is asserted in the suite:

- all 10 hero sheets transferred byte-identical (110 text fields diffed against
  the pre-merge module, zero differences)
- all 12 gameContent.FOES and 8 ADVERSARIES exist in FOES / ADVERSARY_ROSTER at
  the same level, and gain the id/tier/source they lacked
- the 10 quest companions are exactly the COMPANION_CARDS rows carrying a
  `quest`, with identical titles; nameConsistency.test.ts asserts the 10/12 split

The sheet fields are optional because four heroes genuinely have no sheet — the
unreleased Expeditions heroes, whose cards are not public. They carry none of it
rather than a guess, matching how this package treats every unobserved row.
heroes.test.ts asserts a hero has the whole sheet or none of it, and
virtues.test.ts asserts the namespace stays gone.

BREAKING CHANGE: the `gameContent` namespace is removed. gameContent.HEROES ->
HERO_BY_NAME or HERO_BY_ID; .heroes -> HEROES; .FOES -> FOES; .ADVERSARIES ->
ADVERSARY_ROSTER; .COMPANIONS -> COMPANION_CARDS (a superset — the quest ten are
the rows with `quest`); .KINGDOM_VIRTUES/.kingdomVirtues/.Virtue -> the flat
equivalents. `hero.expansion` is gone; it agreed with `hero.source` on all 10
rows, so `source` is the single vocabulary. New: HERO_BY_NAME, and Virtue /
KINGDOM_VIRTUES / kingdomVirtues / KingdomDirection flat from virtues.ts.
